+++ indexer/includes/config.php
@@ -9,8 +9,8 @@ define("NETWORK", $network);
 
 // BTNS Indexer Version
 define("VERSION_MAJOR", 0);
-define("VERSION_MINOR", 11);
-define("VERSION_REVISION",1);
+define("VERSION_MINOR", 12);
+define("VERSION_REVISION",0);
 define("VERSION_STRING", VERSION_MAJOR . '.' . VERSION_MINOR . '.' . VERSION_REVISION);
 
 // TICK constants
@@ -37,9 +37,13 @@ if(NETWORK=='mainnet'){
     // First block with BTNS transaction
     define("FIRST_BLOCK",789742);
 
-    // BTNS Address 
+    // BTNS Addresses
     define('BURN_ADDRESS', "1Muhahahahhahahahahahhahahauxh9QX");
     define('GAS_ADDRESS', "1BTNSGASK5En7rFurDJ79LQ8CVYo2ecLC8");
+
+    // Donation Addresses 
+    define('DONATE_ADDRESS_1', "1BTNSGASK5En7rFurDJ79LQ8CVYo2ecLC8"); // Protocol Development
+    define('DONATE_ADDRESS_2', "1BTNSGASK5En7rFurDJ79LQ8CVYo2ecLC8"); // Community Develoment
 }
 
 // Testnet config
@@ -50,6 +54,10 @@ if(NETWORK=='testnet'){
     // BTNS Address 
     define('BURN_ADDRESS', "mvCounterpartyXXXXXXXXXXXXXXW24Hef");
     define('GAS_ADDRESS', "mvThcDEbeqog2aJ7JNj1FefUPaNdYYGqHt");
+
+    // Donation Addresses 
+    define('DONATE_ADDRESS_1', "mvThcDEbeqog2aJ7JNj1FefUPaNdYYGqHt"); // Protocol Development
+    define('DONATE_ADDRESS_2', "mvThcDEbeqog2aJ7JNj1FefUPaNdYYGqHt"); // Community Develoment
 }
 
 // Database Credentials
@@ -64,10 +72,14 @@ require_once('profiler.php');
 // Rollback code
 require_once('rollback.php');
 
+// Reparse code
+require_once('reparse.php');
+
 // Protocol Changes / Activation blocks
 require_once('protocol_changes.php');
 
 // BTNS Actions
+require_once('actions/address.php');
 require_once('actions/airdrop.php');
 require_once('actions/batch.php');
 require_once('actions/bet.php');