+++ indexer/indexer.php
@@ -24,6 +24,8 @@
  * --testnet    Load data from testnet
  * --block=#    Load data for given block
  * --rollback=# Rollback data to a given block
+ * --reparse    Reparse ALL data 
+ * --reparse=#  Reparse data from a given block
  * --single     Load single block
  ********************************************************************/
 
@@ -31,9 +33,10 @@
 error_reporting(E_ERROR|E_PARSE);
 
 // Parse in any command line args and set basic runtime flags
-$args     = getopt("", array("testnet::", "block::", "single::", "rollback::",));
+$args     = getopt("", array("testnet::", "block::", "single::", "rollback::", "reparse::"));
 $testnet  = (isset($args['testnet'])) ? true : false;
 $single   = (isset($args['single'])) ? true : false;  
+$reparse  = (isset($args['reparse'])) ? ((is_numeric($args['reparse'])) ? $args['reparse'] : true) : false;
 $block    = (is_numeric($args['block'])) ? intval($args['block']) : false;
 $network  = ($testnet) ? 'testnet' : 'mainnet';
 $rollback = (is_numeric($args['rollback'])) ? intval($args['rollback']) : false;
@@ -58,6 +61,10 @@ initDB();
 
 // Create a lock file, and bail if we detect an instance is already running
 createLockFile();
+
+// Define global assoc arrays to track address/ticker changes
+$addresses = [];
+$tickers   = [];
 
 // Handle rollbacks
 if($rollback)
@@ -85,96 +92,21 @@ $lockfile = '/var/tmp/' . $service . '2mysql-' . $network . '.lock';
 if(file_exists($lockfile))
     byeLog("found {$service} parsing a block... exiting");
 
+// Handle reparses
+if($reparse)
+    btnsReparse($current, $reparse);
+
 // Loop through the blocks until we are current
 while($block <= $current){
     $timer = new Profiler();
     print "processing block {$block}...";
 
-    // Lookup any BTNS action broadcasts in this block (anything with bt: or btns: prefix)
-    $sql = "SELECT
-                b.text,
-                b.value as version,
-                t.hash as tx_hash,
-                a.address as source
-            FROM
-                {$dbase}.broadcasts b,
-                {$dbase}.index_transactions t,
-                {$dbase}.index_addresses a
-            WHERE 
-                t.id=b.tx_hash_id AND
-                a.id=b.source_id AND
-                b.block_index='{$block}' AND
-                b.status='valid' AND
-                (b.text LIKE 'bt:%' OR b.text LIKE 'btns:%')
-            ORDER BY b.tx_index ASC";
-    $results = $mysqli->query($sql);
-    if($results){
-        if($results->num_rows){
-            while($row = $results->fetch_assoc()){
-                // Assoc arrays to track address/ticker changes
-                $addresses = array();
-                $tickers   = array();
-                $error     = false;                  
-                $row       = (object) $row;
-                $prefixes  = array('/^bt:/','/^btns:/');
-                $params    = explode('|',preg_replace($prefixes,'',$row->text));
-                $version   = $row->version;  // Project Version
-                $source    = $row->source;   // Source address
+    // Get any broadcast transactions for this block and process them
+    $txs = getBroadcastTransactions($block);
+    foreach($txs as $tx)
+        processTransaction($tx);
 
-                // Create database records and get ids for tx_hash and source address
-                $source_id  = createAddress($row->source);
-                $tx_hash_id = createTransaction($row->tx_hash);
-
-                // Trim whitespace from any PARAMS
-                foreach($params as $idx => $value)
-                    $params[$idx] = trim($value);
-
-                // Extract ACTION from PARAMS
-                $action = strtoupper(array_shift($params)); 
-
-                // Support legacy BTNS format with no VERSION on DEPLOY/MINT/TRANSFER actions (default to VERSION 0)
-                if(in_array($action,array('DEPLOY','MINT','TRANSFER')) && isLegacyBTNSFormat($params))
-                    array_splice($params, 0, 0, 0);
-
-                // Support old BRC20/SRC20 actions 
-                if($action=='TRANSFER') $action = 'SEND';
-                if($action=='DEPLOY')   $action = 'ISSUE';
-
-                // Define basic BTNS transaction data object
-                $data = (object) array(
-                    'ACTION'      => $action,       // Action (ISSUE, MINT, SEND, etc)
-                    'BLOCK_INDEX' => $block,        // Block index 
-                    'SOURCE'      => $row->source,  // Source/Broadcasting address
-                    'TX_HASH'     => $row->tx_hash  // Transaction Hash
-                );
-
-                // Validate Action
-                if(!array_key_exists($action,PROTOCOL_CHANGES))
-                    $error = 'invalid: Unknown ACTION';
-
-                // Verify action is activated (past ACTIVATION_BLOCK)
-                if(!$error && !isEnabled($action, $network, $block))
-                    $error = 'invalid: ACTIVATION_BLOCK';
-
-                // Set action to UNKNOWN if we detect error
-                if($error)
-                    $data->ACTION = $action = 'UNKNOWN';
-
-                // Create a record of this transaction in the transactions table
-                createTxIndex($data);
-
-                // Get tx_index of record using tx_hash
-                $data->TX_INDEX = getTxIndex($data->TX_HASH);
-
-                // Handle processing the specific BTNS ACTION commands
-                btnsAction($action, $params, $data, $error);
-            }
-        }
-    } else {
-        byeLog("Error while trying to lookup BTNS broadcasts");
-    }
-
-    // Create hash of the credits/debits/balances table and create record in `blocks` table
+    // Create record in `blocks` table with hashes of the credits/debits/transactions tables
     createBlock($block);
 
     // Do a sanity check to verify that token supplys match data in credits/debits/balances tables 
@@ -201,4 +133,7 @@ while($block <= $current){
 // Remove the lockfile now that we are done running
 removeLockFile();
 
-print "Total Execution time: " . $runtime->finish() ." seconds\n";
+// Print out information on the total runtime 
+printRuntime($runtime->finish());
+
+?>